Add files via upload
This is a web application made for implementing our dbms project : BLOOD BANK MANAGEMENT SYSTEM.
App.js is the main frontend file, and we have used reactjs in it, App.css is the css file of our frontend. We have used Axios wth reactjs. 
index.js is the main backend file and we have used nodejs, express and vscode's mysql extension in it.
There are two package files, one is the frontend react app and the other is of the server app in nodejs.
Made by: PARTH SHARMA 19BCE0761 and KABIR WADHWA 19BCE2359